+++ src/Client/client_Position.cpp
@@ -1,7 +1,7 @@
 #include <cmath>
 #include "Position.h"
 
-Position::Position(float x, float y): x(x), y(y){}
+Position::Position(int x, int y): x(x), y(y){}
 		
 Position::~Position(){}
 
@@ -9,10 +9,10 @@ bool Position::operator==(const Position& other){
 	return this->x == other.x && this->y == other.y;
 }
 
-float Position::getX() const{
+int Position::getX() const{
 	return this->x;
 }
 
-float Position::getY() const{
+int Position::getY() const{
 	return this->y;
 }
